+++ CommodityTradingAPI/CommodityTradingAPI/Models/TradeMitigation.cs
@@ -7,9 +7,9 @@ public partial class TradeMitigation
 {
     public Guid MitigationId { get; set; }
 
-    public long? SellPointProfit { get; set; }
+    public decimal? SellPointProfit { get; set; }
 
-    public long? SellPointLoss { get; set; }
+    public decimal? SellPointLoss { get; set; }
 
     public virtual ICollection<Trade> Trades { get; set; } = new List<Trade>();
 }
